Adding experiment notes and logs for Alexnet
Although we did not keep track of source code changes for
Alexnet we did take notes on our changes and save the log files.
We actually also saved the last weight checkpoint for every
experiment but unfortunately the files are too large to put
on github. At a latter point we'll try hosting them online
somewhere for reference.